requirements.txt
Django==1.9.7
django-crispy-forms==1.6.0
django-money==0.8
mysqlclient==1.3.7
py-moneyed==0.6.0
